requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-jose[cryptography]   # For JWT
passlib
bcrypt==4.0.0               # For password hashing
python-multipart            # For file uploads
pillow                      # Image processing
google-cloud-storage        # GCS
kafka-python                # Kafka client
redis                       # For rate limiting and caching
SQLAlchemy
psycopg2-binary 
google-cloud-storage        # Google Cloud Storage